Небольшие фиксы

+++ Bullet.h
@@ -5,7 +5,7 @@ struct Bullet : PhysicalObject {
 
     float power;
     float angle;
-    float lifetime;
+    int lifetime;
 
     Bullet(b2World&, const sf::Vector2f&, const sf::Vector2f&, const float&, const float&, sf::Texture*);
